Logic for testing A then B laser tripping
The idea now is to have two scripts running simultaneously.

Script 1 waits for laser A to trip then the script waits for n seconds
for B to trip, if B trips, then +1 to the DB, if B does not trip after n
seconds then restart the loop.

Script 2 does the same thing but waits for B first, then A, and does the
-1 to dB